#2324 FIX Add icon for those checks that cannot be rescheduled
The newest version of Check_MK (when using CMC) detects data that originates
from an agent cache. In such a case a rescheduling of the check is pointless an
the icon for doing that had been removed. In order to reduce confusion amongst
the users a replacement icon has been added instead with an explanatory text.