#2339 FIX Discovery service: Fixed handling of agent / snmp communication errors
When a host could not be contacted via agent or did not receive an SNMP response,
this resulted in a situation where the check was shown as UNKNOWN and broken check
with the hint to send a bug report the us.

This handling has been changed to reporting a WARNING state together with the related
error message as output. The "crashed check" link is not displayed for this case anymore.