#2326 Allow relating Event Console hosts to monitoring hosts now also by alias
In the view of the Event Console events in the GUI Check_MK tries to match the
hosts of the events with the hosts in the monitoring (if they are known) by
either the host name or the IP address. From now on also the hosts' aliases
are being checked.