#2473 FIX cisco_asa_failover: Reworked check to reflect expected primary/secondary states of devices
The check was keeping the current cluster/node states which the devices had during service
discovery as reference state and checks whether or not this state remains and raises an
error when the state changed.

This has been changed now. The state during discovery is not relevant anymore. The check
now looks at the primary/secondary devices and checks whether or not the primary device
is the active one or the secondary has been promoted to be the active one. This should
be clearer for users.