#2540 FIX Process piggy backed data even if host is set to "No Agent"
You need this fix for monitoring a VM just via ESX with no agent in the VM
itself.  You can set this now to <i>No Agent</i> and the services shipped
from the ESX host or vCenter via piggy back are anyway correctly added to
the host. Before this fix the host would just get a PING check.