#2428 FIX "Clustered services for overlapping cluster": Improved rule matching
The matching algorhytm for this rules did not check if the matched
cluster did actually belong to the node.

For example if you had the following clusters
<pre>
ClusterA: Host1, Host2
ClusterB: Host3, Host4
</pre>
And these "Clustered services for overlapping cluster" rules
<pre>
Rule 1) Service1 -> ClusterA
Rule 2) Service1 -> ClusterB
</pre>

If Host3 has a Service1, the rule number one would match and
assign this service to ClusterA, ignoring the fact that this
host is actually a part of ClusterB.

This has been fixed.