#1301 FIX cmciii: Some rewrites of the cmciii Temperature and Power Phase Checks
The Temperature part now uses the default Check_MK Temperature features. This means you can set levels with WATO now.

The Power Phase Check didn't work with Rittal MPSM Devices so it was needed to rewrite it. Cause of this, the name of the Item has changed,
and you have to do a rediscovery.