#2870 dotnet_clrmemory, wmi_cpuload, wmi_webservices, msexch_activesync, msexch_autodiscovery, msexch_availability, msexch_isstore, msexch_owa, msexch_rpcclientaccess: a whole slew of new checks relating to monitoring MS Exchange
All checks are mostly concerned with monitoring Exchange (and other related services) performance,
not with the discovery of failures.
The necessary performance values are automatically generated by the windows agent via wmi if they
are available.
Please note that for many of these there are no levels yet, so they are only informative.