#2825 FIX Parameter overview page no longer raises an exception in certain scenarios
The parameter overview page raised an exception when the host in question never
got a service discovery and had an active check configured.

It also crashed for any "Classical active and passive Monitoring checks" check
with configured rule options, e.g a comment was set or the rule was deactivated.

Conflicts:
	ChangeLog
	modules/automation.py