#2946 FIX LDAP: User created during login does not trigger full user synchronisation anymore
Previously, when a LDAP user logged in for the first time and the user was not created
by an LDAP sync before, it triggered a full LDAP sync which might have created a lot
of other users. This has been changed now to only synchronize this single user based
on the sync configuration.